Added DosingPumpIntervalSetup to features page
When enabled, it overrides the standard dosing pump in PDE generation. So to use all you need to do is enable it and choose a Dosing Pump.

Fixed an error with enabling/disabling PDE controlled features, the features were not being disabled after being enabled when you remove a device/feature.